Moved files - Moved files into a unit folder to seperate them from integration tests.
